refactor: 패키지 구조도 변경
