+++ src/main/java/blackjack/domain/gambler/Gambler.java
@@ -1,6 +1,5 @@
 package blackjack.domain.gambler;
 
-import blackjack.domain.Hands;
 import blackjack.domain.card.Card;
 import java.util.List;
 import java.util.Objects;
